#36 extract clearCart function
